requirements.txt
pydicom>=2.5.0
numpy>=1.25.0
Pillow>=10.0.0
